refactor : model 클래스 domain으로 변경
